*fixed 64bit compile errors
git-svn-id: https://mameplus.svn.sourceforge.net/svnroot/mameplus@4554 e0215e0a-1e64-f442-9f1d-d2f61a7f4648

+++ trunk/mamep/src/emu/uilang.c
@@ -190,10 +190,10 @@ static void load_mmo(int msgcat)
 
 	for (i = 0; i < p->header.num_msg; i++)
 	{
-		p->mmo_index[i].uid = (const unsigned char *)p->mmo_str + (unsigned long)p->mmo_index[i].uid;
-		p->mmo_index[i].ustr = (const unsigned char *)p->mmo_str + (unsigned long)p->mmo_index[i].ustr;
-		p->mmo_index[i].wid = (const unsigned char *)p->mmo_str + (unsigned long)p->mmo_index[i].wid;
-		p->mmo_index[i].wstr = (const unsigned char *)p->mmo_str + (unsigned long)p->mmo_index[i].wstr;
+		p->mmo_index[i].uid = (const unsigned char *)p->mmo_str + (unsigned long long)p->mmo_index[i].uid;
+		p->mmo_index[i].ustr = (const unsigned char *)p->mmo_str + (unsigned long long)p->mmo_index[i].ustr;
+		p->mmo_index[i].wid = (const unsigned char *)p->mmo_str + (unsigned long long)p->mmo_index[i].wid;
+		p->mmo_index[i].wstr = (const unsigned char *)p->mmo_str + (unsigned long long)p->mmo_index[i].wstr;
 	}
 
 	p->status = mmo::MMO_READY;
